Fix casting

+++ aroma/tests/test_integration.py
@@ -50,4 +50,4 @@ def test_integration(skip_integration, nilearn_data):
 
     # Check motion ICs
     mot_ics = np.loadtxt(join(out_path, 'classified_motion_ICs.txt'), delimiter=',')
-    assert np.array_equal(int(mot_ics), np.array([ 1, 2, 3, 4, 6, 7, 12, 13, 20, 24]))
+    assert np.array_equal(mot_ics.astype(int), np.array([ 1, 2, 3, 4, 6, 7, 12, 13, 20, 24]))
